+++ test/CMakeLists.txt
@@ -8,8 +8,8 @@ include_directories(${GMOCK_INCLUDE_DIRS})
 set(GTEST "/usr/src/gtest/libgtest.a" "/usr/src/gmock/libgmock.a")
 
 # Create an executable called runTests with the test file you want to run
-add_executable(runTests tests.cc)
-target_link_libraries(runTests ${GTEST_LIBRARIES} pthread)
-target_link_libraries(runTests ${GTEST})
+add_executable(DummyFsTest dummy-filesystem/dummy-fs-tests.cc)
+target_link_libraries(DummyFsTest ${GTEST_LIBRARIES} pthread)
+target_link_libraries(DummyFsTest ${GTEST})
 include_directories(../dummy-filesystem)
 link_directories(/home/vagrant/rdfs/build/dummy-filesystem)
