--- test/turtle.h
@@ -1,14 +1,0 @@
-//src: https://github.com/biicode/docs/blob/master/c%2B%2B/examples/gmock.rst
-
-#pragma once
-class Turtle {
-  public: 
-    virtual ~Turtle() {}
-    virtual void PenUp() = 0;
-    virtual void PenDown() = 0;
-    virtual void Forward(int distance) = 0;
-    virtual void Turn(int degrees) = 0;
-    virtual void GoTo(int x, int y) = 0;
-    virtual int GetX() const = 0;
-    virtual int GetY() const = 0;
-};
